create helper function to parse application resource requirements move image extraction func to helper file
